+++ src/main/java/com/gigeroa/main/java/age_of_certantitos/servicios/CalculadorDistancia.java
@@ -5,8 +5,8 @@ import com.gigeroa.main.java.age_of_certantitos.servicios.CalculadorDistancia;
 
 public abstract class CalculadorDistancia{
 
-	public static Integer calcularDistancia(Posicion posicion1, Posicion posicion2) {
-		Integer calculo = posicion1.getEjeX()-posicion2.getEjeX();
+	public static Double calcularDistancia(Posicion posicion1, Posicion posicion2) {
+		Double calculo = posicion1.getEjeX()-posicion2.getEjeX();
 		//TODO Falta la posicion en Y
 		if (calculo < 0){
 			return -calculo;
